+++ src/main/java/io/bookwright/api/model/semaphore/Template.java
@@ -15,4 +15,6 @@ public record Template(
     String arguments,
     @JsonProperty("allow_override_args_in_task") boolean allowOverrideArgsInTask,
     @JsonProperty("survey_vars") List<SurveyVariable> surveyVariables,
-    @JsonProperty("task_params") AnsibleTemplateParameters taskParameters) {}
+    @JsonProperty("task_params") AnsibleTemplateParameters taskParameters,
+    @JsonProperty("runner_tag") String runnerTag,
+    @JsonProperty("allow_parallel_tasks") boolean allowParallelTasks) {}